Remove name option when listing jobs
The acceptance tests were failing.
In attempting to make them pass, I concluded they couldn't ever
work.

As it was, the acceptance tests were 404ing, on URLs like:
`/job/test/api/json`. This was because there is no job called `test`. If
it was changed to search for `this.jobName`, then it would make a
successful request. But the response type would be different - it wouldn't
have the `jobs` array field that the following code looks for.

I can't find anyway to filter the `/api/json` route by jobname.
As such, I think it's better to just get rid of the option, as it can't
be supported.